refactor(rest api): moved rest api to get latest transactions to its own file and added get transaction by uuid
