+++ server/controllers/collectives.js
@@ -1,7 +1,7 @@
 /**
  * Dependencies.
  */
-import { get, pick, omit } from 'lodash';
+import { pick, omit } from 'lodash';
 import async from 'async';
 import {
   defaultHostCollective,
@@ -19,52 +19,8 @@ import errors from '../lib/errors';
 import debugLib from 'debug';
 import config from 'config';
 import prependHttp from 'prepend-http';
-import * as utils from '../graphql/v1/utils';
 
 const { Activity, Notification, Collective, ConnectedAccount, User } = models;
-
-const allTransactionsQuery = `
-  query allTransactions($collectiveSlug: String!, $limit: Int, $offset: Int, $type: String, $includeVirtualCards: Boolean ) {
-    allTransactions(collectiveSlug: $collectiveSlug, limit: $limit, offset: $offset, type: $type, includeVirtualCards: $includeVirtualCards) {
-      id
-      uuid
-      type
-      amount
-      currency
-      hostCurrency
-      hostCurrencyFxRate
-      hostFeeInHostCurrency
-      platformFeeInHostCurrency
-      paymentProcessorFeeInHostCurrency
-      netAmountInCollectiveCurrency
-      createdAt
-      updatedAt
-      host {
-        id
-        slug
-      }
-      createdByUser {
-        id
-        email
-      }
-      fromCollective {
-        id
-        slug
-        name
-        image
-      }
-      collective {
-        id
-        slug
-        name
-        image
-      }
-      paymentMethod {
-        id
-      }
-    }
-  }
-`;
 
 const _addUserToCollective = (collective, user, options) => {
   const checkIfCollectiveHasHost = () => {
@@ -702,22 +658,4 @@ export const getTransactions = (req, res, next) => {
       );
     })
     .catch(next);
-};
-
-/**
- * Get array of all transactions of a collective given its slug
- */
-export const getCollectiveTransactions = async (req, res) => {
-  try {
-    const args = req.query;
-    args.collectiveSlug = get(req, 'params.collectiveSlug');
-    const response = await utils.graphqlQuery(allTransactionsQuery, req.query);
-    if (response.errors) {
-      throw new Error(response.errors[0]);
-    }
-    const result = get(response, 'data.allTransactions', []);
-    res.send({ result });
-  } catch (error) {
-    res.status(400).send({ error: error.toString() });
-  }
 };
